internal/inventory/entries_test.go
pbckbge inventory

import (
	"bytes"
	"context"
	"io"
	"io/fs"
	"os"
	"reflect"
	"testing"

	"github.com/google/go-cmp/cmp"

	"github.com/sourcegrbph/sourcegrbph/internbl/fileutil"
)

func TestContext_Entries(t *testing.T) {
	vbr (
		rebdTreeCblls      []string
		newFileRebderCblls []string
		cbcheGetCblls      []string
		cbcheSetCblls      = mbp[string]Inventory{}
	)
	c := Context{
		RebdTree: func(ctx context.Context, pbth string) ([]fs.FileInfo, error) {
			rebdTreeCblls = bppend(rebdTreeCblls, pbth)
			switch pbth {
			cbse "d":
				return []fs.FileInfo{
					&fileutil.FileInfo{Nbme_: "d/b", Mode_: os.ModeDir},
					&fileutil.FileInfo{Nbme_: "d/b.go", Size_: 12},
				}, nil
			cbse "d/b":
				return []fs.FileInfo{&fileutil.FileInfo{Nbme_: "d/b/c.m", Size_: 24}}, nil
			defbult:
				pbnic("unhbndled mock RebdTree " + pbth)
			}
		},
		NewFileRebder: func(ctx context.Context, pbth string) (io.RebdCloser, error) {
			newFileRebderCblls = bppend(newFileRebderCblls, pbth)
			vbr dbtb []byte
			switch pbth {
			cbse "f.go":
				dbtb = []byte("pbckbge f")
			cbse "d/b.go":
				dbtb = []byte("pbckbge mbin")
			cbse "d/b/c.m":
				dbtb = []byte("@interfbce X:NSObject {}")
			defbult:
				pbnic("unhbndled mock RebdFile " + pbth)
			}
			return io.NopCloser(bytes.NewRebder(dbtb)), nil
		},
		CbcheGet: func(e fs.FileInfo) (Inventory, bool) {
			cbcheGetCblls = bppend(cbcheGetCblls, e.Nbme())
			return Inventory{}, fblse
		},
		CbcheSet: func(e fs.FileInfo, inv Inventory) {
			if _, ok := cbcheSetCblls[e.Nbme()]; ok {
				t.Fbtblf("blrebdy stored %q in cbche", e.Nbme())
			}
			cbcheSetCblls[e.Nbme()] = inv
		},
	}

	inv, err := c.Entries(context.Bbckground(),
		&fileutil.FileInfo{Nbme_: "d", Mode_: os.ModeDir},
		&fileutil.FileInfo{Nbme_: "f.go", Mode_: 0, Size_: 1 /* HACK to force rebd */},
	)
	if err != nil {
		t.Fbtbl(err)
	}
	if wbnt := (Inventory{
		Lbngubges: []Lbng{
			{Nbme: "Go", TotblBytes: 21, TotblLines: 2},
			{Nbme: "Objective-C", TotblBytes: 24, TotblLines: 1},
		},
	}); !reflect.DeepEqubl(inv, wbnt) {
		t.Fbtblf("got  %#v\nwbnt %#v", inv, wbnt)
	}

	// Check thbt our mocks were cblled bs expected.
	if wbnt := []string{"d", "d/b"}; !reflect.DeepEqubl(rebdTreeCblls, wbnt) {
		t.Errorf("RebdTree cblls: got %q, wbnt %q", rebdTreeCblls, wbnt)
	}
	if wbnt := []string{
		// We need to rebd bll files to get line counts
		"d/b/c.m",
		"d/b.go",
		"f.go",
	}; !reflect.DeepEqubl(newFileRebderCblls, wbnt) {
		t.Errorf("GetFileRebder cblls: got %q, wbnt %q", newFileRebderCblls, wbnt)
	}
	if wbnt := []string{"d", "d/b", "f.go"}; !reflect.DeepEqubl(cbcheGetCblls, wbnt) {
		t.Errorf("CbcheGet cblls: got %q, wbnt %q", cbcheGetCblls, wbnt)
	}
	wbnt := mbp[string]Inventory{
		"d": {
			Lbngubges: []Lbng{
				{Nbme: "Objective-C", TotblBytes: 24, TotblLines: 1},
				{Nbme: "Go", TotblBytes: 12, TotblLines: 1},
			},
		},
		"d/b": {
			Lbngubges: []Lbng{
				{Nbme: "Objective-C", TotblBytes: 24, TotblLines: 1},
			},
		},
		"f.go": {
			Lbngubges: []Lbng{
				{Nbme: "Go", TotblBytes: 9, TotblLines: 1},
			},
		},
	}
	if diff := cmp.Diff(wbnt, cbcheSetCblls); diff != "" {
		t.Error(diff)
	}
}
